Incorporate code from StackTraceFilter at its single point of call and remove it
